requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
pytz==2024.1
aiofiles==23.2.1
aiosqlite==0.19.0
python-docx==1.1.2
Pillow==10.4.0
